Remove the exclusive-or language around DT or ACPI
Early in EBBR discussions the decision was made that firmware should not
provide both DT and ACPI at the same time. The reasoning had been that
we didn't want to encourage 'hybrid' approaches where the OS tries to
consume both DT and ACPI descriptions.

However, this fear has not so-far born out, and feedback has been that
requiring a boot-time switch to select either DT or ACPI is a support
issue for OS vendors. They would rather see ACPI (or DT) always turned
on if it is an option and the OS can choose to use it or not.

Thoughts? Can the exclusive or be relaxed here?

Signed-off-by: Grant Likely <grant.likely@arm.com>